fix: test to use expected
