+++ src/components/Icon/Icon.spec.tsx
@@ -1,19 +1,20 @@
-import React from 'react';
-
 import { render } from '@testing-library/react';
 
 import Icon from './Icon';
 
 describe('Icon component', () => {
   it('renders without crashing', () => {
-    render(<Icon icon="alert" />);
+    const { container } = render(<Icon icon="alert" />);
+    expect(container).toBeInTheDocument();
   });
 
   it('renders with optional size', () => {
-    render(<Icon icon="alert" size={20} />);
+    const { container } = render(<Icon icon="alert" size={20} />);
+    expect(container).toBeInTheDocument();
   });
 
   it('renders with optional width and height', () => {
-    render(<Icon height={30} icon="alert" width={30} />);
+    const { container } = render(<Icon height={30} icon="alert" width={30} />);
+    expect(container).toBeInTheDocument();
   });
 });
