+++ src/components/Icon/Icon.types.ts
@@ -1,0 +1,10 @@
+import { SVGProps } from 'react';
+
+import { icons } from './data';
+
+export type TIconsType = keyof typeof icons;
+
+export interface IIconProps extends SVGProps<SVGSVGElement> {
+  icon: TIconsType;
+  size?: number;
+}
