+++ src/components/Icon/data.ts
@@ -1,0 +1,27 @@
+import Alert from './icons/alert.svg?react';
+import Check from './icons/check.svg?react';
+import Close from './icons/close.svg?react';
+import Hamburguer from './icons/hamburguer.svg?react';
+import LeftArrow from './icons/left-arrow.svg?react';
+import Mag from './icons/mag.svg?react';
+import Minus from './icons/minus.svg?react';
+import Plus from './icons/plus.svg?react';
+import RightArrow from './icons/right-arrow.svg?react';
+import SmallCircleFilled from './icons/small-circle-filled.svg?react';
+import StarFilled from './icons/star-filled.svg?react';
+import Star from './icons/star.svg?react';
+
+export const icons = {
+  alert: Alert,
+  check: Check,
+  close: Close,
+  hamburguer: Hamburguer,
+  'left-arrow': LeftArrow,
+  mag: Mag,
+  minus: Minus,
+  plus: Plus,
+  'right-arrow': RightArrow,
+  'small-circle-filled': SmallCircleFilled,
+  star: Star,
+  'star-filled': StarFilled,
+};
