+++ src/components/Modal/Modal.tsx
@@ -4,11 +4,9 @@ import { createPortal } from 'react-dom';
 import classNames from 'classnames';
 import { AnimatePresence, motion } from 'framer-motion';
 
-import Icon from '~components/Icon/Icon';
-
 import scss from './Modal.module.scss';
 
-import { type TModalProps } from './Modal.types';
+import type { TModalProps } from './Modal.types';
 
 function Modal(props: TModalProps) {
   useEffect(() => {
@@ -22,14 +20,14 @@ function Modal(props: TModalProps) {
 
   return createPortal(
     <AnimatePresence>
-      {props.isOpen && (
+      {props.isOpen ? (
         <motion.div
-          data-testid="portal"
           animate={{ opacity: 1 }}
-          exit={{ opacity: 0 }}
-          transition={{ duration: 0.3 }}
           className={classNames(scss.container, props.className)}
+          data-testid="portal"
+          exit={{ opacity: 0 }}
           onClick={props.onClickOutside}
+          transition={{ duration: 0.3 }}
         >
           <div
             className={scss.modalContent}
@@ -38,9 +36,7 @@ function Modal(props: TModalProps) {
             <div className={scss.modalHeader}>
               <h2>{props.title}</h2>
               <div className={scss.headerButtons}>
-                <button onClick={props.onClickOutside}>
-                  {<Icon icon="emote" size="large" />}
-                </button>
+                <button onClick={props.onClickOutside} />
               </div>
             </div>
             <div>
@@ -49,7 +45,7 @@ function Modal(props: TModalProps) {
             <footer className={scss.modalFooter}>{props.footer}</footer>
           </div>
         </motion.div>
-      )}
+      ) : null}
     </AnimatePresence>,
     document.body
   );
